fix: election voting count logic
