+++ src/domains/cluster_actors/mod.rs
@@ -11,7 +11,6 @@ pub mod commands;
 
 pub mod consensus;
 
-pub mod election_state;
 pub mod heartbeats;
 pub mod replication;
